Refactored UiUpdater to delegate to CommandExecutor which can be run on both the client and the server. They both operate on StringBuffers so this works well. Obviously if the underlying file on the server differs from what the client percieves, then there could be errors, but given the same inputs, the frontend and backend will produce the same output.
Also created the NotificationSystem so that info and error messages properly show up on the screen, and added a counter for commands to see how many commands there are in the queue.

The editor now properly displays the file on separate lines, but currently doesn't show blank lines. Also, the only commands available are typing and backspace, so the carat can't be moved to edit existing text. The commands sent to the backend are completely uncompressed and therefore super inefficient (one Command object per character), but it should be simple enough to add a compression step to turn the individual commands into multiple.